+++ oneadapter/src/main/java/com/idanatz/oneadapter/internal/paging/OneScrollListener.kt
@@ -4,7 +4,9 @@ import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import com.idanatz.oneadapter.internal.utils.Logger
-import com.idanatz.oneadapter.internal.utils.extensions.findLastVisibleItemPosition
+import com.idanatz.oneadapter.internal.utils.extensions.findLastItemIndex
+import com.idanatz.oneadapter.internal.utils.extensions.findLastVisibleItemIndex
+import com.idanatz.oneadapter.internal.utils.extensions.isScrolling
 
 internal class OneScrollListener (
         private val layoutManager: RecyclerView.LayoutManager,
@@ -28,7 +30,7 @@ internal class OneScrollListener (
 
     // This happens many times a second during a scroll, so be wary of the code you place here.
     override fun onScrolled(view: RecyclerView, dx: Int, dy: Int) {
-        if (!isUserScrolled(view)) {
+        if (!view.isScrolling) {
             return
         }
 
@@ -64,13 +66,12 @@ internal class OneScrollListener (
         previousTotalItemCount = 0
     }
 
-    private fun isUserScrolled(view: RecyclerView) = view.scrollState != RecyclerView.SCROLL_STATE_IDLE
-
     private fun evaluateLoadingState(): LoadingState {
-        val lastItemIndex = layoutManager.itemCount - 1
-        val lastVisibleItemIndex = layoutManager.findLastVisibleItemPosition()
+        val lastItemIndex = layoutManager.findLastItemIndex()
+        val lastVisibleItemIndex = layoutManager.findLastVisibleItemIndex()
 
         return when {
+			lastItemIndex == RecyclerView.NO_POSITION || lastVisibleItemIndex == RecyclerView.NO_POSITION -> LoadingState.Normal
             isLoadingFinished(lastItemIndex) -> LoadingState.FinishLoading
             shouldStartLoading(lastVisibleItemIndex, lastItemIndex) -> LoadingState.LoadingStarted
             loading -> LoadingState.MidLoading